WIP: add missing closing HTML tag
No, the closing tag is not located in any of the included files.
Heck, the bug makes it through all the way to the client – you
can see it in the HTML hosted at <https://3mdeb.com/bug-bounty/>.
(Compare that with <https://3mdeb.com/careers/>, which has the
"Contact Us" `<section>` as the last direct child of `<main>`.

And btw, is there really no tooling to catch things like that?