WIP: fix broken spacing
See the open positions listing at <https://3mdeb.com/praca/>

This is how the EN version of the document seems to deal with the problem.
However, the issue does not occur in the "Do oddziału w Osiu poszukujemy:"
section for some reason.

Speaking of non-breaking spaces, the entire repo is filled with them!
But not as HTML entities like here, but rather as U+00A0 characters
inserted throughout the text. It's probably because many XKB layouts
have it mapped under Alt+Space, which is quite easy to hit on accident.

(there's a ton of other accidental whitespace too)

Also, should all of this really be taken care of manually? Why is the
green text all applied by hand, with Markdown?
And are the EN and PL versions both kept in sync/up-to-date in any way?